Split encoding::types into encoding-types crate
This allows for creation of alternative non-WHATWG encodings that use the same
interface as encodings defined in this crate without pulling in all the
tables and encodings.

This commit does not introduce any breaking changes; all the types previously defined in
encoding::types are reexported.

Fixes #81.
