+++ Cargo.toml
@@ -14,6 +14,10 @@ license = "MIT"
 [lib]
 name = "encoding"
 
+[dependencies.encoding-types]
+version = "0.2"
+path = "src/types"
+
 # version policy for index tables:
 # - major: addition or deletion of index tables
 # - minor: any content changes to index tables, numbered by the date
@@ -44,4 +48,3 @@ path = "src/index/tradchinese"
 
 [dev-dependencies]
 getopts = "*" # for examples
-
